test: stop double-deleting indices in task list tests
using a little state to track if we've already deleted them or not as we
have multiple places where we attempt to do so and it's a little unclear
whether all places are used in all tests or not.

this removes some errors in the tests logs, but also avoids performing
extra ES/OS calls - which should slightly speed things up.