fix: validate schema field differences when mapping should not be dynamic
If we are not expecting an index's mapping to be dynamic, but it turns
out to be then we should ensure we do validate the fields to ensure they
are the correct type.

As ES/OS will auto-create indexes with dynamic mappings it would be
quite a common thing for clients to accidentally have indexes setup this
way.  This can lead to unusual bugs and potentially hard to diagnose
bugs  e.g. if a field we expect to be "keyword" is actually "text" then
term queries will behave in a very different way.

As it's quite likely we have self-managed clients with this happening
I've added a flag to be able to turn this behaviour off (as a
workaround).

I've also updated the logging for when we are ignoring these changes to
WARN as it's probably quite an important thing to be aware of when
looking into client issues.